Completed preponderance of refactor, including moving engine constructor out of global context.
