+++ Library/Source/NativeEnginePlugin.h
@@ -1,0 +1,6 @@
+#include <napi/napi.h>
+
+namespace babylon
+{
+    Napi::Function CreateNativeEngineConstructor(Napi::Env&);
+}
