+++ TestApp/Scripts/experience.js
@@ -46,7 +46,7 @@ function CreateInputHandling(scene) {
     });
 }
 
-var nativeEngine = new NativeEngineImpl();
+var nativeEngine = new _native.Engine();
 
 var engine = new BABYLON.NativeEngine();
 var scene = new BABYLON.Scene(engine);
